requirements.txt
fastapi[all]
flake8==6.0.0
fastapi-users[sqlalchemy]
aiosqlite
sqladmin[full]
asyncpg==0.29.0
easyocr
pillow==10.2.0